% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_rmse}
\alias{compute_rmse}
\title{Compute the Relative Integrated Mean Squared Error}
\usage{
compute_rmse(true, fit)
}
\arguments{
\item{true}{Discrete evaluations of the true function}

\item{fit}{Discrete evaluations of the estimated function}
}
\value{
The R-MISE value
}
\description{
\code{compute_rmse()} computes the Relative Integrated Mean Sqaured Error (R-MISE) between
a reference function and its estimate, evaluated at the same input points.
}
