% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crMM_NoWarp.R
\name{crMM_NoWarp}
\alias{crMM_NoWarp}
\title{MCMC sampler for functional mixed membership functional without warping}
\usage{
crMM_NoWarp(
  num_it,
  burnin = 0.2,
  t,
  y,
  p,
  degree_shape = 3,
  intercept_shape = FALSE,
  a_e,
  b_e,
  a_c,
  b_c,
  a_l,
  b_l,
  rescale_pi = TRUE,
  tuning_pi = 1000,
  alpha,
  label1 = NULL,
  label2 = NULL,
  wantPAF = FALSE,
  gamma1_init = NULL,
  gamma2_init = NULL,
  lambda1_init = 0.1,
  lambda2_init = 0.1,
  var_c_init = 0.1,
  var_e_init = 1
)
}
\arguments{
\item{num_it}{The number of MCMC iterations (after burn-in).}

\item{burnin}{The proportion of the iterations eliminated for burn-in. The default is \code{0.2}.
The first \code{burnin * num_it} iterations are discarded, and the sampler then runs for another \code{num_it}
iterations from which samples are recorded.}

\item{t}{The time points at which the data is observed.}

\item{y}{A matrix containing the functional observations. Each row corresponds to one observation.}

\item{p}{The number of inner knots used to define the B-splines for the common shape functions.
The sampler assumes equally spaced knots between the first and last value of \code{t}.}

\item{degree_shape}{Degree of the piecewise polynomial of B-splines for the common shape functions.
The default is \code{3} for cubic splines.}

\item{intercept_shape}{If \code{TRUE} an intercept is included in the B-spline basis for the common
shape functions. The default is \code{FALSE}.}

\item{a_e, a_c, a_l}{Shape parameters of the Inverse Gamma priors.}

\item{b_e, b_c, b_l}{Rate parameters of the Inverse Gamma priors.}

\item{rescale_pi}{If \code{TRUE}, the values of the mixed membership vector pi are rescaled such that
at least one observation belongs entirely to each cluster. Default is \code{TRUE}.}

\item{tuning_pi}{Scaling for the shape parameter of the Metropolis proposal Dirichlet distribution
for sampling of pi. Default is \code{1000}.}

\item{alpha}{Shape parameter for the Dirichlet prior for pi.}

\item{label1, label2}{Vector containing the indices of observations labelled as belonging to
feature 1 and feature 2 respectively. Default is \code{NULL}, in which case observations are not labelled.}

\item{wantPAF}{If \code{TRUE}, the sampler records the PAF value at each iteration (only relevant for
the EEG case study). Default is \code{FALSE}.}

\item{gamma1_init, gamma2_init}{Initial value the B-spline coefficients for the common shape functions.
Default is \code{NULL}, in which case the sampler obtains coefficients based on the extreme observations.}

\item{lambda1_init, lambda2_init, var_c_init, var_e_init}{Initial values for prior variances.
Default is \code{0.1}, except for \code{var_e_init} which is \code{1}.}
}
\value{
An object of class \code{crMM_Obj}, consisting of a list with the following items:
\itemize{
\item \code{gamma1}: a matrix with \code{num_it} rows of the posterior samples for the B-spline coefficient
for the shape associated with the first feature.
\item \code{gamma2}: a matrix with \code{num_it} rows of the posterior samples for the B-spline coefficient
for the shape associated with the second feature.
\item \code{c}: a matrix with \code{num_it} rows of the posterior samples for the individual intercepts. The number
of columns corresponds to the number of subjects, that is, \code{nrow(y)}.
\item \code{variance}: a matrix with \code{num_it} rows of the posterior samples for the variance parameters.
The four columns, in order, correspond to the variances for the first and second feature B-spline
coefficients, the intercepts, and the error term.
\item \code{pi1}: a matrix with \code{num_it}rows of the posterior samples of the mixed membership component for the
first feature. The number of columns corresponds to the number of subjects, that is,
the number of rows in \code{y}.
\item \code{fit_sample}: matrix with \code{num_it} rows of the posterior samples of individual fits. The first
\code{lenght(t)} columns give the fit for the first observation, the second \code{lenght(t)} columns give
the fit for the second observation, and so on for all the data.
\item \code{fit1}, \code{fit2}: matrices with \code{nrow(y)} rows corresponding to the ergodic first and second
sample moment of the fit at points \code{t} for each of the observations.
}

Additionally, if \code{wantPAF} is \code{TRUE} then the list also includes a matrix with \code{num_it} rows of the
posterior samples of the Peak Alpha Frequency.
}
\description{
\code{crMM_NoWarp()} runs a Metropolis-within-Gibbs sampler for the analysis of functional data under the two
feature mixed membership model assumption. This function does not account for phase variation, see
\code{\link[=crMM_Warp]{crMM_Warp()}} and \code{\link[=crMM_WarpReg]{crMM_WarpReg()}} for samplers that include estimation of time-transformation functions.
}
