% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crMM_WarpReg.R
\name{crMM_WarpReg}
\alias{crMM_WarpReg}
\title{MCMC sampler for curve registration with regression for functional mixed membership models}
\usage{
crMM_WarpReg(
  num_it,
  burnin = 0.2,
  t,
  y,
  X,
  p,
  degree_shape = 3,
  intercept_shape = FALSE,
  inc_rho = TRUE,
  rho_init = 0.5,
  h,
  degree_tt = 3,
  intercept_tt = FALSE,
  a_e,
  b_e,
  a_c,
  b_c,
  a_l,
  b_l,
  a_phi,
  b_phi,
  rescale_pi = TRUE,
  B0,
  V0,
  B_init = B0,
  tuning_pi = 1000,
  alpha,
  label1 = NULL,
  label2 = NULL,
  wantPAF = FALSE,
  gamma1_init = NULL,
  gamma2_init = NULL,
  lambda1_init = 0.1,
  lambda2_init = 0.1,
  var_c_init = 0.1,
  var_e_init = 1,
  var_phi_init = 1,
  process_id = NULL
)
}
\arguments{
\item{num_it}{The number of MCMC iterations (after burn-in).}

\item{burnin}{If a number between 0 and 1, the proportion of the iterations eliminated for burn-in.
The first \code{burnin * num_it} iterations are discarded, and the sampler then runs for another \code{num_it}
iterations from which samples are recorded. If an integer, the number of iterations to discard, before
running for another \code{num_it} iterations from which samples are recorded. The default is \code{0.2}.}

\item{t}{The time points at which the data is observed.}

\item{y}{A matrix containing the functional observations. Each row corresponds to one observation.}

\item{X}{Matrix of covariates for time-transformation regression, where each row corresponds to covariates for
one observation.}

\item{p}{The number of inner knots used to define the B-splines for the common shape functions.
The sampler assumes equally spaced knots between the first and last value of \code{t}.}

\item{degree_shape}{Degree of the piecewise polynomial of B-splines for the common shape functions.
The default is \code{3} for cubic splines.}

\item{intercept_shape}{If \code{TRUE} an intercept is included in the B-spline basis for the common
shape functions. The default is \code{FALSE}.}

\item{inc_rho}{If \code{TRUE}, a parameter that rescales the time transformation functions for the second
feature towards the identity. Default is \code{TRUE}.}

\item{rho_init}{Initial value for the time-transformation rescaling parameter. Default is \code{0.5}.}

\item{h}{The number of inner knots used to define the B-splines for the time-transformation functions.
The sampler assumes equally spaced knots between the first and last value of \code{t}.}

\item{degree_tt}{Degree of the piecewise polynomial of B-splines for the time-transformation functions.
The default is \code{3} for cubic splines.}

\item{intercept_tt}{If \code{TRUE} an intercept is included in the B-spline basis for the time-transformation
functions. The default is \code{FALSE}.}

\item{a_e, a_c, a_l, a_phi}{Shape parameters of the Inverse Gamma priors.}

\item{b_e, b_c, b_l, b_phi}{Rate parameters of the Inverse Gamma priors.}

\item{rescale_pi}{If \code{TRUE}, the values of the mixed membership vector pi are rescaled such that
at least one observation belongs entirely to each cluster. Default is \code{TRUE}.}

\item{B0}{Prior mean matrix for the regression coefficient.}

\item{V0}{Prior row covariance matrix for the regression coefficient.}

\item{B_init}{Initial value for the regression coefficient. Default is \code{B0}.

@returns An object of class \code{crMM_Obj}, consisting of a list with the following items:
\itemize{
\item \code{gamma1}: a matrix with \code{num_it} rows of the posterior samples for the B-spline coefficient
for the shape associated with the first feature.
\item \code{gamma2}: a matrix with \code{num_it} rows of the posterior samples for the B-spline coefficient
for the shape associated with the second feature.
\item \code{c}: a matrix with \code{num_it} rows of the posterior samples for the individual intercepts. The number
of columns corresponds to the number of subjects, that is, \code{nrow(y)}.
\item \code{variance}: a matrix with \code{num_it} rows of the posterior samples for the variance parameters.
The five columns, in order, correspond to the variances for the first and second feature B-spline
coefficients, the intercepts, the error term, and the time-transformation spline coefficients.
\item \code{phi}: a matrix with \code{num_it} rows of the posterior samples for the B-spline coefficients for the
individual time-transformation functions. The coefficients across individuals are stacked in order into
one vector corresponding to a row at each iteration.
\item \code{pi1}: a matrix with \code{num_it} rows of the posterior samples of the mixed membership component for the
first feature. The number of columns corresponds to the number of subjects, that is,
the number of rows in \code{y}.
\item \code{B}: a matrix with \code{num_it} rows of the posterior samples of the time-transformation regression
coefficient matrix. The entries of the matrix are column-stacked into a vector, the columns correspond
to the entries.
\item \code{fit_sample}: matrix with \code{num_it} rows of the posterior samples of individual fits. The first
\code{length(t)} columns give the fit for the first observation, the second \code{length(t)} columns give
the fit for the second observation, and so on for all the data.
\item \code{registered_fit}: matrix with \code{num_it} rows of the posterior samples of individual aligned fits. The first
\code{length(t)} columns give the fit for the first observation, the second \code{length(t)} columns give
the fit for the second observation, and so on for all the data.
\item \code{stochastic_time}: matrix with \code{num_it} rows of the posterior samples of the individual stochastic
(aligned) time. The first \code{length(t)} columns give the times for the first observation, the second
\code{length(t)} columns give the times for the second observation, and so on for all the data.
\item \code{fit1}, \code{fit2}: matrices with \code{nrow(y)} rows corresponding to the ergodic first and second
sample moment of the fit at points \code{t} for each of the observations.
\item \code{loglik} : vector with \code{num_it} rows of the full model loglikelihood at each iteration.
}

Additionally, if \code{wantPAF} is \code{TRUE} then the list also includes a matrix \code{PAF} with \code{num_it} rows of the
posterior samples of the Peak Alpha Frequency.

Also, if \code{inc_rho} is \code{TRUE}, the list includes a matrix \code{rho} with \code{num_it} rows of the posterior samples
for the time-transformation rescaling parameter. Also, \code{stochastic_time2}, a matrix with \code{num_it} rows
of the posterior samples of the individual stochastic (aligned) time for the scaled second feature is
provided. The first \code{length(t)} columns give the times for the first observation, the second
\code{length(t)} columns give the times for the second observation, and so on for all the data.}

\item{tuning_pi}{Scaling for the shape parameter of the Metropolis proposal Dirichlet distribution
for sampling of pi. Default is \code{1000}.}

\item{alpha}{Shape parameter for the Dirichlet prior for pi.}

\item{label1, label2}{Vector containing the indices of observations labelled as belonging to
feature 1 and feature 2 respectively. Default is \code{NULL}, in which case observations are not labelled.}

\item{wantPAF}{If \code{TRUE}, the sampler records the PAF value at each iteration (only relevant for
the EEG case study). Default is \code{FALSE}.}

\item{gamma1_init, gamma2_init}{Initial value the B-spline coefficients for the common shape functions.
Default is \code{NULL}, in which case the sampler obtains coefficients based on the extreme observations.}

\item{lambda1_init, lambda2_init, var_c_init, var_e_init, var_phi_init}{Initial values for prior variances.
Default is \code{0.1}, except for \code{var_e_init} and \code{var_phi_init} which is \code{1}.}

\item{process_id}{If provided, gives an ID to the iteration for parallel processing. Default is \code{NULL}.}
}
\description{
\code{crMM_WarpReg()} runs a Metropolis-within-Gibbs sampler for the analysis of functional data under the two-feature
mixed membership model assumption. This function incorporates curve registration to account for phase variation.
It differs from \code{\link[=crMM_Warp]{crMM_Warp()}} in that it includes covariate information in the estimation of the
time-transformation functions through linear regression.
}
