% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_data}
\alias{plot_data}
\title{Plot of the data}
\usage{
plot_data(
  t,
  y,
  indices = 1:nrow(y),
  xlab = "x",
  ylab = "y",
  title = "",
  viridis_option = "C",
  color = "grey"
)
}
\arguments{
\item{t}{Time points for the x axis. Either a vector that is common for all observations, or a matrix for which
each row corresponds to an observation.}

\item{y}{Matrix of data. Each row corresponds to one observation.}

\item{indices}{Vector indicating which observations to plot. Default is \code{1:nrow(y)}, indicating that all obersvations
should be plotted.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"y"}.}

\item{title}{Title for the figure. Default is \code{""}.}

\item{viridis_option}{String indicating color map to use for \code{scale_color_viridis_d()}. Default is \code{"C"}.
If \code{NULL}, plots in a single color rather than with the viridis gradient.}

\item{color}{Color to plot in if \code{viridis_option} is \code{NULL}. Default is \code{"grey"}.}
}
\value{
A ggplot figure
}
\description{
\code{plot_data()} plots the rows of the matrix of observations.
}
