% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_feature}
\alias{plot_feature}
\title{Plot of posterior feature shape}
\usage{
plot_feature(
  t,
  shape,
  quantile_low,
  quantile_high,
  color = "black",
  alpha = 0.2,
  xlab = "x",
  ylab = "f",
  title = "",
  background_ind = NULL,
  y = NULL
)
}
\arguments{
\item{t}{Time points for the x axis.}

\item{shape}{Posterior central summary of the shape function.}

\item{quantile_low}{Posterior lower quantile of the shape function.}

\item{quantile_high}{Posterior upper quantile of the shape function.}

\item{color}{Line color for the shape function. Default is \code{"black"}.}

\item{alpha}{Opacity of the color for the quantile ribbon. Default is \code{0.2}.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \verb{"f"'}.}

\item{title}{Title for the figure. Default is \code{""}.}

\item{background_ind}{Indices of observations to plot. Default is \code{NULL}, in which case no observations are
displayed with the posterior shape.}

\item{y}{Matrix of data. Each row corresponds to one observation. Default is \code{NULL}, for the case when no observations
are displayed with the posterior shape.}
}
\value{
A ggplot figure
}
\description{
\code{plot_feature()} plots the posterior shape function for the desired feature. The posterior quantiles of interest
are also represented.
}
