% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot of posterior fit}
\usage{
plot_fit(
  t,
  y,
  fit,
  indices = sample(1:nrow(y), 5),
  xlab = "x",
  ylab = "y",
  title = "",
  line_colors = c("red", "blue", "green", "purple", "orange")
)
}
\arguments{
\item{t}{Time points for the x axis. Either a vector that is common for all observations, or a matrix for which
each row corresponds to an observation.}

\item{y}{Matrix of data. Each row corresponds to one observation.}

\item{fit}{Matrix of fit. Each row corresponds to one observation.}

\item{indices}{Vector indicating observations for which to plot the fit. Default is \code{sample(1:nrow(y), 5)}.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"y"}.}

\item{title}{Title for the figure. Default is \code{""}.}

\item{line_colors}{Vectors of colors for the lines of fit. Default assumes 5 lines are plotted, and is
\code{c("red", "blue", "green", "purple", "orange")}.}
}
\value{
A ggplot figure
}
\description{
\code{plot_fit()} plots the posterior fit for the desired observations, overlain on the true data.
}
