% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_identifiability}
\alias{plot_identifiability}
\title{Plot shape functions from different samples}
\usage{
plot_identifiability(
  eval_t,
  shape,
  colors = c("red", "blue", "green", "orange", "purple"),
  xlab = "x",
  ylab = "f",
  title = ""
)
}
\arguments{
\item{eval_t}{Time points for the horizontal axis.}

\item{shape}{List whose elements are the values of the shape function for different simulations.}

\item{colors}{Vectors of colors for the line of each shape. Default is
\verb{("red", "blue", "green", "orange", "purple")}, assuming 5 lines are plotted.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"f"}.}

\item{title}{Title for the figure. Default is \code{""}.}
}
\value{
A ggplot figure
}
\description{
\code{plot_identifiability()} plots the posterior shape function for the desired feature obtained through
different methods or simulations. It focuses on the identifiability, for plots with confidence bands
and legends see \code{plot_multiple_feature()}.
}
