% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_member_boxplot}
\alias{plot_member_boxplot}
\title{Plot a boxplot of individual mixed membership components}
\usage{
plot_member_boxplot(
  pi,
  group,
  colors = c("cornflowerblue", "gold1"),
  group_labels = c("0", "1"),
  xlab = "Designation",
  ylab = "Value",
  title = ""
)
}
\arguments{
\item{pi}{Values of the mixed membership componenet for each individual.}

\item{group}{Vector of 0 and 1s indicating group belonging for each individual.}

\item{colors}{Vector of two colors, where each corresponds to the color of the boxplot for one group.
Default is \code{c("cornflowerblue", "gold")}.}

\item{group_labels}{Labels to use to denote each group. Default is \code{c("0", "1")}.}

\item{xlab}{Label for the horizontal axis. Default is \code{"Designation"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"Value"}.}

\item{title}{Title for the figure. Default is \code{""}.}
}
\value{
A ggplot figure
}
\description{
\code{plot_member_boxplot()} plots a boxplot for the membership to the provided feature, separated
by group. In the case study, the group corresponds to the clinical designation.
}
