% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_multiple_feature}
\alias{plot_multiple_feature}
\title{Plot of multiple posterior feature shapes}
\usage{
plot_multiple_feature(
  eval_t,
  shape,
  quantile_low,
  quantile_high,
  colors,
  alpha = 0.2,
  xlab = "x",
  ylab = "f",
  title = "",
  background_ind = NULL,
  t = NULL,
  y = NULL,
  legend_position = c(0.8, 0.8)
)
}
\arguments{
\item{eval_t}{Time points for the horizontal axis.}

\item{shape}{List whose elements are the posterior central summary of the shape function.}

\item{quantile_low}{List whose elements are the posterior lower quantile of the shape function.}

\item{quantile_high}{List whose elements are the posterior upper quantile of the shape function.}

\item{colors}{Vector of colors for the line and ribbon for each different shape.}

\item{alpha}{Opacity of the color for the quantile ribbon. Default is \code{0.2}.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"f"}.}

\item{title}{Title for the figure. Default is \code{""}.}

\item{background_ind}{Indices of observations to plot. Default is \code{NULL}, in which case no observations
are displayed with the posterior shape.}

\item{t}{Time points at which data is evaluated. Default is \code{NULL}, for the case when no observations
are displayed with the posterior shape.}

\item{y}{Matrix of data. Each row corresponds to one observation. Default is \code{NULL}, for the case when
no observations are displayed with the posterior shape.}

\item{legend_position}{Vector of coordinates for the location of the legend. Default is \code{c(0.8, 0.8)}.}
}
\value{
A ggplot figure
}
\description{
\code{plot_multiple_feature()} plots the posterior shape function for the desired feature obtained through
different methods or simulations. The posterior quantiles of interest are also represented.
}
