% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_posterior_hist}
\alias{plot_posterior_hist}
\title{Plot a histogram and density estimate for a sample}
\usage{
plot_posterior_hist(
  sample,
  colors = c("gold", "darkblue"),
  xlab = "x",
  ylab = "Density",
  title = "",
  binwidth = 0.03
)
}
\arguments{
\item{sample}{Sample of values for which the distribution is of interest.}

\item{colors}{Vector of two colors. The first one is for the fill color of the histogram bins, the
second one is the line color for the density. Default is \code{c("gold", "darkblue")}.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"Density"}.}

\item{title}{Title for the figure. Default is \code{""}.}

\item{binwidth}{Width of histogram bins. Default is \code{0.03}.}
}
\value{
A ggplot figure
}
\description{
\code{plot_posterior_hist()} returns the histogram and density estimate for a sample. In particular,
it allows to illustrate the posterior distribution of parameters or quantities of interest.
}
