% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_tt}
\alias{plot_tt}
\title{Plot of posterior time transformation functions}
\usage{
plot_tt(
  eval_t,
  tt_functions,
  group = NULL,
  covariate = NULL,
  group_colors = c("gold1", "cornflowerblue"),
  group_labels = c("0", "1"),
  covariate_gradient = c("yellow", "blue"),
  viridis_option = "C",
  xlab = "Physical Time",
  ylab = "Stochastic Time",
  title = "",
  covlab = "Age"
)
}
\arguments{
\item{eval_t}{Time points for the horizontal axis.}

\item{tt_functions}{Matrix of time transformation functions, where each row corresponds to one function.}

\item{group}{Vector of 0 and 1s indicating group belonging for each function. Default is \code{NULL}, for when the group
is not provided and all time transformation functions are plotted together.}

\item{covariate}{Vector of covariate values associated with each function. Default is \code{NULL}, for when covariate
value is not provided.}

\item{group_colors}{Line colors for each group. Default is \code{c("gold1", "cornflowerblue")}.}

\item{group_labels}{Labels to use to denote each group. Default is \code{c("0", "1")}.}

\item{covariate_gradient}{Two colors determining the gradient for the covariate values. Default is
\code{c("yellow", "blue")}.}

\item{viridis_option}{String indicating color map to use for \code{scale_color_viridis_d()}. Default is \code{"C"}.}

\item{xlab}{Label for the horizontal axis. Default is \code{"Physical Time"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"Stochastic Time"}.}

\item{title}{Title for the figure. Default is \code{""}.}

\item{covlab}{Title for the legend for covariate values. Default is \code{"Age"}.}
}
\value{
A ggplot figure.
}
\description{
\code{plot_tt()} plots time-transformation functions. Options allow to plot all together or split into two groups, as
well as color based on covariate value.
}
