% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{posterior_regression}
\alias{posterior_regression}
\title{Posterior estimates for the time-transformation regression parameter}
\usage{
posterior_regression(
  crMM_samples,
  moment = "mean",
  h,
  degree = 3,
  intercept = FALSE,
  CaseStudy = FALSE,
  ages = NULL,
  t_boundary = c(0, 1),
  eval_points = 100
)
}
\arguments{
\item{crMM_samples}{An object of class \code{crMM_Obj}.}

\item{moment}{Posterior moment of interest. Options are: \code{"mean"}, \code{"median"}, \code{"sd"} or values
between 0 and 1 for quantiles. The default is \code{"mean"}.}

\item{h}{The number of inner knots used to define the B-spline for the time-transformation functions.}

\item{degree}{Degree of the piecewise cubic polynomial of B-splines for the time-transformation
functions. The default is \code{3} for cubic splines.}

\item{intercept}{If \code{TRUE}, an intercept is included in the B-spline basis for the time-transformation
functions. The default is \code{FALSE}.}

\item{CaseStudy}{Logical that indicates whether the sample consists of the analysis of the EEG case study.
The default is \code{FALSE}.}

\item{ages}{When \code{CaseStudy = TRUE}, the age values at which to evaluate the regression mean for the
time-transformation functions. Default is \code{NULL}, for the general case.}

\item{t_boundary}{A vector with the first and last observation time point. Default is \code{c(0,1)}.}

\item{eval_points}{When \code{CaseStudy = TRUE}, number of points at which to evaluate the time-transformation
functions using B-spines. Default is \code{100}.}
}
\value{
When \code{CaseStudy = FALSE}, returns the posterior sample summary for the regression coefficient.

When \code{CaseStudy = TRUE}, returns a list containing:
\itemize{
\item \code{B}: the posterior sample summary of the regression coefficient
\item \code{means0}: a matrix with the regression mean function evaluated at different ages, for TD individuals
\item \code{means1}: a matrix with the regression mean function evaluated at different ages, for ASD individuals
}
}
\description{
\code{posterior_regression()} obtains the posterior summary statistics for the time-transformation regression
parameter. It calculates the desired summary statistics / moments of the posterior MCMC samples for the
regression parameter. If the sample provided considers the analysis of the EEG case study, the
mean time-transformation functions associated with inputted ages, and stratified by clinical designation,
are also fitted.
}
