% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{posterior_tt}
\alias{posterior_tt}
\title{Posterior estimates for time-transformation functions}
\usage{
posterior_tt(
  crMM_samples,
  h,
  degree = 3,
  intercept = FALSE,
  moment = "mean",
  t_boundary,
  eval_points
)
}
\arguments{
\item{crMM_samples}{An object of class \code{crMM_Obj}.}

\item{h}{The number of inner knots used to define the B-spline for the time-transformation functions.}

\item{degree}{Degree of the piecewise cubic polynomial of B-splines for the time-transformation
functions. The default is \code{3} for cubic splines.}

\item{intercept}{If \code{TRUE}, an intercept is included in the B-spline basis for the time-transformation
functions. The default is \code{FALSE}.}

\item{moment}{Posterior moment of interest. Options are: \code{"mean"}, \code{"median"}, \code{"sd"} or values
between 0 and 1 for quantiles. The default is \code{"mean"}.}

\item{t_boundary}{A vector with the first and last observation time point. Default is \code{c(0,1)}.}

\item{eval_points}{When \code{CaseStudy = TRUE}, number of points at which to evaluate the time-transformation
functions using B-spines. Default is \code{100}.}
}
\value{
When there is no rescaling of the time-transformations in the second feature, returns a list containing:
\itemize{
\item \code{phi} : a matrix for which each row corresponds to the posterior sample summary for an individual time-transformation
B-spline coefficient
\item \code{tt_functions}: a matrix for which each row corresponds to the posterior sample summary for an individual
time-transformation function
\item \code{stochastic_time}: a matrix for which each row corresponds to the posterior sample summary of the individual
stochastic time points
}

If there is rescaling, the list additionally contains:
\itemize{
\item \code{rho}: the posterior sample summary of the time-transformation rescaling parameter
\item \code{tt_functions2} : a matrix for which each row corresponds to the posterior sample summary for an individual
time-transformation function associated with the second feature
\item \code{stochastic_time}: a matrix for which each row corresponds to the posterior sample summary of the individual
stochastic time points associated with the second feature
}
}
\description{
\code{posterior_tt()} obtains the desired posterior summary statistics for quantities related to the time-
transformation functions. If the sample provided does not including rescaling for the second
feature, the posterior summary is given for the individual time-transformation B-spline coefficients,
the individual time-transformation functions, and the stochastic time points. When rescaling for the second feature,
posterior summaries of the rescaling parameter, as well as the time-transformation functions and stochastic time
points associated with feature 2 are also computed.
Note that this function is designed for two-feature mixed membership models with optional time-transformation
rescaling of the second feature, as used in the EEG case study (it does not generalize to K features).
}
