% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{r_mise}
\alias{r_mise}
\title{Relative Mean Integrated Square Error}
\usage{
r_mise(true, fit)
}
\arguments{
\item{true}{A vector or matrix with observed true values of a function. If a matrix, each row corresponds
to a different function.}

\item{fit}{A vector or matrix with obersved fitted values of a function. If a matrix, each row corresponds
to a different function.}
}
\value{
The R-MISE value. If \code{fit} is a matrix, then the output is a vector of R-MISE values.
}
\description{
\code{r_mise()} computes the relative mean integrated square error (R-MISE) between two functions.
}
\details{
The user may provide a single reference \code{true} function, and multiple fitted functions \code{fit}, as long
as the coordinates at which they are observed coincide. The opposite is not possible: one cannot provide
multiple \code{true} functions, but provide a single \code{fit} function.
}
