% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_data}
\alias{plot_data}
\title{Plot of the data}
\usage{
plot_data(t, y, indices = 1:nrow(y), xlab = "x", ylab = "y", title = "")
}
\arguments{
\item{t}{Time points for the x axis. Either a vector that is common for all observations, or a matrix for which
each row corresponds to an observation.}

\item{y}{Matrix of data. Each row corresponds to one observation.}

\item{indices}{Vector indicating which observations to plot. Default is \code{1:nrow(y)}, indicating that all obersvations
should be plotted.}

\item{xlab}{Label for the horizontal axis. Default is \code{"x"}.}

\item{ylab}{Label for the vertical axis. Default is \code{"y"}.}

\item{title}{Title for the figure. Default is \code{""}.}
}
\value{
A gglpot figure
}
\description{
\code{plot_data()} plots the rows of the matrix of observations.
}
