% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summaries.R
\name{posterior_shape}
\alias{posterior_shape}
\title{Posterior estimates of the shape functions}
\usage{
posterior_shape(
  crMM_samples,
  t,
  p,
  degree = 3,
  intercept = FALSE,
  moments = c("mean"),
  eval_points = 1000
)
}
\arguments{
\item{crMM_samples}{An object of class \code{crMM_Obj}.}

\item{t}{The time points at which the data is observed.}

\item{p}{The number of inner knots used to define the B-splines for the common shape functions.}

\item{degree}{Degree of the piecewise polynomial of B-splines for the common shape functions.
The default is \code{3} for cubic splines.}

\item{intercept}{If \code{TRUE} an intercept is included in the B-spline basis for the common
shape functions. The default is \code{FALSE}.}

\item{moments}{Vector with the posterior moments of interest. Options that can be included are:
\code{"mean"}, \code{"median"}, \code{"sd"}, or values between 0 and 1 for quantiles. The default is \code{c("mean")}.}

\item{eval_points}{Number of points at which to evaluated the function. Default is \code{1000}.}
}
\value{
A list containing a list for each of the inputted moments, that consists of:
\itemize{
\item \code{gamma} : a list of \code{K} elements containing the posterior summaries of the samples for
the B-spline coefficient associated with each of the \code{K} features.
\item \code{f} : a list containing the evaluated functions associated with each features using the
estimate from \code{gamma}.
}
}
\description{
\code{posterior_shape()} obtains posterior summaries for the shape functions. It calculates the desired summary
statistic/ moments of the posterior MCMC samples for the B-spline coefficients of the K features,
and fits the associated posterior function.
}
